+++ CHANGELOG.md
@@ -3,6 +3,16 @@
 ## Unreleased
 
 
+## 1.1.0 [2025/11/08]
+### Added
+- feat: Manage images on comments
+- feat: Manage image form display
+- feat: Add admin management panel
+
+### Fixed
+- fix: CSS style on idea card
+
+
 ## 1.0.1 [2025/11/03]
 ### Fixed
 - fix: Default values